LPS-92702 Move test classes of com.liferay.portal.verify package to module
